Remove emulator, Fix usings, Separate Extensions to flies
