+++ src/Monik.Client.Base/MonikClient.cs
@@ -1,7 +1,6 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Threading.Tasks;
+﻿using System.Collections.Concurrent;
 using System.Threading;
+using System.Threading.Tasks;
 using Monik.Common;
 
 namespace Monik.Client
