+++ src/Monik.Client.Base/MonikConsole.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Text;
 
 namespace Monik.Common
 {
