+++ src/Monik.Common/Bootstrapper.cs
@@ -1,8 +1,4 @@
-﻿#if (EMULATOR)
-using Monik.Service.Test;
-#endif
-
-using System;
+﻿using System;
 using Autofac;
 using Gerakul.FastSql.Common;
 using Gerakul.FastSql.SqlServer;
@@ -48,18 +44,11 @@ namespace Monik.Service
 
             container.Resolve<IMessageProcessor>().OnStart();
             container.Resolve<IMessagePump>().OnStart();
-
-#if (EMULATOR)
-            container.Resolve<MessageEmulator>().OnStart();
-#endif
         }
 
         // Raise at NancyHostHolder.Stop() when service shutdown
         public void OnApplicationStop()
         {
-#if (EMULATOR)
-            Singleton.Resolve<MessageEmulator>().OnStop();
-#endif
             Singleton.Resolve<IMonik>().OnStop();
             Singleton.Resolve<IMessagePump>().OnStop();
             Singleton.Resolve<IMessageProcessor>().OnStop();
@@ -76,13 +65,8 @@ namespace Monik.Service
 
             existingContainer.RegisterSingleton<IMonikServiceSettings, MonikServiceSettings>();
 
-#if (EMULATOR)
-            // Stub have internal state !
-            existingContainer.RegisterSingleton<IRepository, RepositoryStub>();
-#else
             existingContainer.RegisterInstance<ContextProvider, SqlContextProvider>(SqlContextProvider.DefaultInstance);
             existingContainer.RegisterImplementation<IRepository, Repository>();
-#endif
 
             existingContainer.RegisterSingleton<IMonik, MonikEmbedded>();
 
@@ -97,10 +81,6 @@ namespace Monik.Service
             existingContainer.RegisterImplementation<IMetricObject, MetricObject>();
 
             existingContainer.Update(builder => builder.Register(c => existingContainer));
-
-#if (EMULATOR)
-            existingContainer.Update(builder => builder.RegisterType<MessageEmulator>());
-#endif
         }
 
         protected override void ConfigureRequestContainer(ILifetimeScope container, NancyContext context)
@@ -119,60 +99,5 @@ namespace Monik.Service
             NancyInternalConfiguration.WithOverrides(c => c.Serializers = new [] {typeof(JsonNetSerializer)});
     }
 
-    public static class LifeTimeExtension
-    {
-        public static void RegisterSingleton<TInterface, TImplementation>(this ILifetimeScope container)
-        {
-            container.Update(builder => builder
-                .RegisterType<TImplementation>()
-                .As<TInterface>()
-                .SingleInstance());
-        }
-
-        public static void RegisterMultiplySingleton<TI1, TI2, TImplementation>(this ILifetimeScope container)
-        {
-            container.Update(builder => builder
-                .RegisterType<TImplementation>()
-                .As<TI1>()
-                .As<TI2>()
-                .SingleInstance());
-        }
-
-        public static void RegisterImplementation<TInterface, TImplementation>(this ILifetimeScope container)
-        {
-            container.Update(builder => builder
-                .RegisterType<TImplementation>()
-                .As<TInterface>());
-        }
-
-        public static void RegisterInstance<TInterface, TImplementation>(this ILifetimeScope container, TImplementation aInstance)
-        {
-            container.Update(builder => builder
-                .Register(x => aInstance)
-                .As<TInterface>());
-        }
-
-        public static void RegisterSingleInstance<TInterface, TImplementation>(this ILifetimeScope container, TImplementation aInstance)
-        {
-            container.Update(builder => builder
-                .Register(x => aInstance)
-                .As<TInterface>()
-                .SingleInstance());
-        }
-
-        public static void RegisterNamedSingleton<TInterface, TImplementation>(this ILifetimeScope container, string name)
-        {
-            container.Update(builder => builder
-                .RegisterType<TImplementation>()
-                .Named<TInterface>(name)
-                .SingleInstance());
-        }
-
-        public static void RegisterNamedImplementation<TInterface, TImplementation>(this ILifetimeScope container, string name)
-        {
-            container.Update(builder => builder
-                .RegisterType<TImplementation>()
-                .Named<TInterface>(name));
-        }
-    }//end of class
+    //end of class
 }
