+++ src/Monik.Common/Caches/CacheBase.cs
@@ -1,7 +1,7 @@
-﻿using Monik.Common;
-using System.Collections.Concurrent;
+﻿using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
+using Monik.Common;
 
 namespace Monik.Service
 {
