+++ src/Monik.Common/Caches/CacheLog.cs
@@ -1,7 +1,7 @@
-﻿using Monik.Common;
-using System.Collections.Concurrent;
+﻿using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using Monik.Common;
 
 namespace Monik.Service
 {
