+++ src/Monik.Common/Caches/CacheMetric.cs
@@ -1,9 +1,9 @@
-﻿using Autofac;
-using Monik.Common;
-using System;
+﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
+using Autofac;
+using Monik.Common;
 
 namespace Monik.Service
 {
