+++ src/Monik.Common/Metrics/MetricObject.cs
@@ -1,7 +1,7 @@
-﻿using Monik.Common;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Monik.Common;
 
 namespace Monik.Service
 {
