+++ src/Monik.Common/Processing/MessagePump.cs
@@ -1,9 +1,9 @@
-﻿using Monik.Common;
-using System;
+﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Monik.Common;
 
 namespace Monik.Service
 {
