+++ src/Monik.Common/Processing/MonikEmbedded.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
+﻿using System.Collections.Concurrent;
 using Autofac;
 using Monik.Common;
