+++ src/Monik.Common/Queues/AzureActiveQueue.cs
@@ -1,5 +1,5 @@
-﻿using Monik.Common;
-using System;
+﻿using System;
+using Monik.Common;
 
 namespace Monik.Service
 {
