+++ src/Monik.Common/Settings/MonikServiceSettings.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Configuration;
 
 namespace Monik.Service
